Preparation for 0.1.0-alpha.3 (5)

+++ scirs2-cluster/examples/kmeans2_demo.rs
@@ -83,7 +83,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
         let mut correct = 0;
         for i in 0..true_labels.len() {
             // Simple accuracy assuming cluster ordering matches
-            if true_labels[i] == predicted_labels[i] {
+            if true_labels[i] == predicted_labels[i] as i32 {
                 correct += 1;
             }
         }
